Allow compute services to be initialised from an existing service ID (#125)
* Allow compute services to be initialised from an existing service ID

* Ensure we select the ideal service version to mutate during compute init.

* Update compute init service-id flag description

+++ pkg/compute/init.go
@@ -52,6 +52,7 @@ var (
 // InitCommand initializes a Compute@Edge project package on the local machine.
 type InitCommand struct {
 	common.Base
+	serviceID   string
 	name        string
 	description string
 	author      string
@@ -67,6 +68,7 @@ func NewInitCommand(parent common.Registerer, globals *config.Data) *InitCommand
 	var c InitCommand
 	c.Globals = globals
 	c.CmdClause = parent.Command("init", "Initialize a new Compute@Edge package locally")
+	c.CmdClause.Flag("service-id", "Existing service ID to use. By default, this command creates a new service").Short('s').StringVar(&c.serviceID)
 	c.CmdClause.Flag("name", "Name of package, defaulting to directory name of the --path destination").Short('n').StringVar(&c.name)
 	c.CmdClause.Flag("description", "Description of the package").Short('d').StringVar(&c.description)
 	c.CmdClause.Flag("author", "Author of the package").Short('a').StringVar(&c.author)
@@ -107,6 +109,23 @@ func (c *InitCommand) Exec(in io.Reader, out io.Writer) (err error) {
 
 	undoStack := common.NewUndoStack()
 	defer func() { undoStack.RunIfError(out, err) }()
+
+	var (
+		service *fastly.Service
+		version int
+	)
+
+	if c.serviceID != "" {
+		service, err = c.Globals.Client.GetService(&fastly.GetServiceInput{
+			ID: c.serviceID,
+		})
+		if err != nil {
+			return fmt.Errorf("error fetching service details: %w", err)
+		}
+
+		c.name = service.Name
+		c.description = service.Comment
+	}
 
 	if c.path == "" {
 		fmt.Fprintf(progress, "--path not specified, using current directory\n")
@@ -206,25 +225,55 @@ func (c *InitCommand) Exec(in io.Reader, out io.Writer) (err error) {
 		progress = text.NewQuietProgress(out)
 	}
 
-	progress.Step("Creating service...")
-	service, err := c.Globals.Client.CreateService(&fastly.CreateServiceInput{
-		Name:    c.name,
-		Type:    "wasm",
-		Comment: c.description,
-	})
-	if err != nil {
-		return fmt.Errorf("error creating service: %w", err)
-	}
-	undoStack.Push(func() error {
-		return c.Globals.Client.DeleteService(&fastly.DeleteServiceInput{
-			ID: service.ID,
+	// If we have an existing service, get the ideal version.
+	// Otherwise create a new service and set version to 1.
+	if c.serviceID != "" {
+		versions, err := c.Globals.Client.ListVersions(&fastly.ListVersionsInput{
+			Service: c.serviceID,
 		})
-	})
+		if err != nil {
+			return fmt.Errorf("error listing service versions: %w", err)
+		}
+
+		v, err := getLatestIdealVersion(versions)
+		if err != nil {
+			return fmt.Errorf("error finding latest service version")
+		}
+
+		if v.Active || v.Locked {
+			progress.Step("Cloning latest version...")
+			v, err = c.Globals.Client.CloneVersion(&fastly.CloneVersionInput{
+				Service: c.serviceID,
+				Version: v.Number,
+			})
+			if err != nil {
+				return fmt.Errorf("error cloning latest service version: %w", err)
+			}
+		}
+
+		version = v.Number
+	} else {
+		progress.Step("Creating service...")
+		service, err = c.Globals.Client.CreateService(&fastly.CreateServiceInput{
+			Name:    c.name,
+			Type:    "wasm",
+			Comment: c.description,
+		})
+		if err != nil {
+			return fmt.Errorf("error creating service: %w", err)
+		}
+		version = 1
+		undoStack.Push(func() error {
+			return c.Globals.Client.DeleteService(&fastly.DeleteServiceInput{
+				ID: service.ID,
+			})
+		})
+	}
 
 	progress.Step("Creating domain...")
 	_, err = c.Globals.Client.CreateDomain(&fastly.CreateDomainInput{
 		Service: service.ID,
-		Version: 1,
+		Version: version,
 		Name:    c.domain,
 	})
 	if err != nil {
@@ -233,7 +282,7 @@ func (c *InitCommand) Exec(in io.Reader, out io.Writer) (err error) {
 	undoStack.Push(func() error {
 		return c.Globals.Client.DeleteDomain(&fastly.DeleteDomainInput{
 			Service: service.ID,
-			Version: 1,
+			Version: version,
 			Name:    c.domain,
 		})
 	})
@@ -241,7 +290,7 @@ func (c *InitCommand) Exec(in io.Reader, out io.Writer) (err error) {
 	progress.Step("Creating backend...")
 	_, err = c.Globals.Client.CreateBackend(&fastly.CreateBackendInput{
 		Service: service.ID,
-		Version: 1,
+		Version: version,
 		Name:    c.backend,
 		Address: c.backend,
 	})
@@ -251,7 +300,7 @@ func (c *InitCommand) Exec(in io.Reader, out io.Writer) (err error) {
 	undoStack.Push(func() error {
 		return c.Globals.Client.DeleteBackend(&fastly.DeleteBackendInput{
 			Service: service.ID,
-			Version: 1,
+			Version: version,
 			Name:    c.backend,
 		})
 	})
